refactor: remove duplicated folder
