+++ src/app/app.component.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AppShellComponent } from './app-shelll/app-shell/app-shell.component';
+import { AppShellComponent } from './app-shell/app-shell.component';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
